@@ -22,7 +22,8 @@
     "__postinstall": "electron-builder install-app-deps"
   },
   "engines": {
-    "node": "12.16.3"
+    "node": "12.16.3",
+    "npm": ">= 7.0.0"
   },
   "main": "background.js",
   "dependencies": {
@@ -109,6 +110,7 @@
   "devDependencies": {
     "@mdi/font": "^5.0.45",
     "@mdi/js": "^5.0.45",
+    "@types/jest": "^24.9.1",
     "@types/pouchdb": "^6.4.0",
     "@types/pouchdb-find": "^6.3.5",
     "@vue/cli-plugin-babel": "^4.4.6",
@@ -133,7 +135,7 @@
     "node-sass": "^4.13.1",
     "sass": "^1.24.3",
     "sass-loader": "^8.0.0",
-    "typescript": "~3.8.3",
+    "typescript": "^4.2.4",
     "vee-validate": "^3.3.7",
     "vue-chartist": "^2.2.1",
     "vue-cli-plugin-i18n": "^0.6.0",
